improve api_key_info fetching
When the access_mask or key_type is not set, they're fetched automatically as soon as they're needed, but guard clauses that check key_type were causing this kind of information to be fetched twice.

Instead get_key_type and get_access_mask will hand off to fetch_key_info which sets both as soon as one is needed, preventing redundant http requests.